update to v1.10.0

+++ models/vm_rebuild_params.go
@@ -57,7 +57,7 @@ type VMRebuildParams struct {
 	MaxBandwidthPolicy *VMDiskIoRestrictType `json:"max_bandwidth_policy,omitempty"`
 
 	// max iops
-	MaxIops *int32 `json:"max_iops,omitempty"`
+	MaxIops *int64 `json:"max_iops,omitempty"`
 
 	// max iops policy
 	MaxIopsPolicy *VMDiskIoRestrictType `json:"max_iops_policy,omitempty"`
